refactor(notifications): make recipient resolver channel-agnostic
Replace InMemoryNotificationRecipientResolver, which hardcoded knowledge of the
Telegram channel, with a CompositeNotificationRecipientResolver that aggregates
profile-gated NotificationDestinationProvider beans:

- LoggingNotificationDestinationProvider (always active, default channel).
- TelegramNotificationDestinationProvider (only @Profile("telegram-notifications")).

Boundary hardening achieved:
- When the Telegram profile is OFF, no Telegram chat-id property is read and no
  TelegramNotificationDestination is ever produced; the listener no longer logs
  the spurious 'no sender available for channel TELEGRAM' warning that could
  occur when chat-ids were configured but the sender bean was absent.
- Adding a new channel (email, SMS, ...) is now a zero-change task on the
  resolver: ship one DestinationProvider + one Sender, both gated by the same
  profile.
- Watchlists module is unaffected and remains channel-agnostic.

All 57 tests pass (up from 55; +2 from Phase 2 module verification expansion).
